Create README.md
his project involves creating an online code editor that runs in browsers, offering complete Integrated Development Environment (IDE) functionalities. The project is divided into two main components: the Backend API and the Frontend Editor. The Backend API is a server-side component that receives code and language as input, executes the code on the server, and returns the output. The Frontend Editor allows users to choose a programming language, edit and modify the code, and then make a POST request to the Backend API to execute the code and display the output on the website. This project not only serves as a valuable skill-building exercise but also has the potential to evolve into a startup, offering a free platform for coding interviews and collaborative coding sessions.